Reorganized metrics
